requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
sentence-transformers
openai
tiktoken
python-dotenv
